+++ src/main/java/org/openwebnet4j/USBGateway.java
@@ -17,6 +17,7 @@ package org.openwebnet4j;
 import java.nio.ByteBuffer;
 
 import org.eclipse.jdt.annotation.NonNull;
+import org.eclipse.jdt.annotation.Nullable;
 import org.openwebnet4j.communication.OWNException;
 import org.openwebnet4j.communication.Response;
 import org.openwebnet4j.communication.USBConnector;
@@ -31,7 +32,7 @@ import org.slf4j.LoggerFactory;
 /**
  * Class {@link USBGateway} to connect to ZigBee USB Gateways using {@link USBConnector}
  *
- * @author M. Valla - Initial contribution
+ * @author M. Valla - Initial contribution. Added SPI and SerialPortProvider injection.
  */
 public class USBGateway extends OpenGateway {
 
@@ -42,15 +43,28 @@ public class USBGateway extends OpenGateway {
     private int receivedProducts = 0; // number of products returned from PRODUCT INFORMATION
     // requests during a device discovery
 
-    // FIXME -SPI- change to private
-    public SerialPortProvider serialPortProvider;
+    private SerialPortProvider serialPortProvider;
 
     public USBGateway(String serialPortName) {
         this.serialPortName = serialPortName;
     }
 
+    /**
+     * Sets a {@link SerialPortProvider} to use to obtain available serial ports
+     *
+     * @param provider the {@link SerialPortProvider} to set
+     */
     public void setSerialPortProvider(@NonNull SerialPortProvider provider) {
         serialPortProvider = provider;
+    }
+
+    /**
+     * Gets the current {@link SerialPortProvider}
+     *
+     * @return {@link SerialPortProvider} currently set
+     */
+    public @Nullable SerialPortProvider getSerialPortProvider() {
+        return serialPortProvider;
     }
 
     /**
@@ -68,7 +82,7 @@ public class USBGateway extends OpenGateway {
         tmpConn.setSerialPortProvider(serialPortProvider);
         connector = tmpConn;
         logger.info("##USB## Init USB ({})...", serialPortName);
-        logger.debug("##USB## CONNECTOR -- SerialPortProvider = {}...", tmpConn.serialPortProvider);
+        logger.debug("##USB## CONNECTOR -- SerialPortProvider = {}...", tmpConn.getSerialPortProvider());
 
     }